Created application system that will abstract threads.
My current plan of action is as follows:
	A recorder thread that manages recording data weekly and tracks how long you are on a window.
	A ewmh thread to poll for when a new window is clicked.
	An alert thread that will spawn a window to notify you as unproductive.

+++ src/application/process.rs
@@ -18,7 +18,7 @@ pub struct Process {
 impl Process {
     unsafe fn set_name(&mut self, ptr : *const c_char){
         for i in 0..strlen(ptr) {
-            self.name.push(*ptr.offset(i as isize) as u8 as char);
+            self.name.push(*ptr.offset(i as isize) as char);
         }
     }
